agents/scene_cut/requirements.txt
google-cloud-storage
firebase-admin
functions-framework # Only needed if deploying as a Google Cloud Function HTTP trigger
